+++ src/components/Notification/index.tsx
@@ -1,11 +1,7 @@
 import React from 'react';
 import Style from './style';
 
-type NotificationProps = {
-  message: string;
-}
-
-const Notification = ({ message }: Readonly<NotificationProps>): JSX.Element =>
-  <Style>{message}</Style>;
+const Notification = ({ message, type }: Readonly<NotificationType>): JSX.Element =>
+  <Style type={type}>{message}</Style>;
 
 export default Notification;